+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 3.404
 
-
+fix sql install violation
 
 
 --------------------