+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-3.404
+3.405
 
+--------------------
+3.404
 fix sql install violation
 Help to generate password